Extract methods to separate class
